Day07

+++ src/main/kotlin/Day.kt
@@ -2,6 +2,6 @@ interface Day {
 
     val fileName: String
 
-    fun solvePart1(input: String): Int = 0
-    fun solvePart2(input: String): Int = 0
+    fun solvePart1(input: String): Long = 0
+    fun solvePart2(input: String): Long = 0
 }
