+++ src/main/kotlin/Day01.kt
@@ -4,7 +4,7 @@ class Day01 : Day {
 
     override val fileName: String = "Day01"
 
-    override fun solvePart1(input: String): Int {
+    override fun solvePart1(input: String): Long {
         val inputLines = input.asLines()
         val (left, right) = makeLists(inputLines)
         return left.sorted()
@@ -12,19 +12,19 @@ class Day01 : Day {
             .sumOf { (it.first - it.second).absoluteValue }
     }
 
-    override fun solvePart2(input: String): Int {
+    override fun solvePart2(input: String): Long {
         val inputLines = input.asLines()
         val (left, right) = makeLists(inputLines)
         val elementsFrequencies = right.groupingBy { it }.eachCount()
         return left.sumOf { it * elementsFrequencies.getOrDefault(it, 0) }
     }
 
-    private fun makeLists(input: List<String>): Pair<MutableList<Int>, MutableList<Int>> {
-        val left = mutableListOf<Int>()
-        val right = mutableListOf<Int>()
+    private fun makeLists(input: List<String>): Pair<MutableList<Long>, MutableList<Long>> {
+        val left = mutableListOf<Long>()
+        val right = mutableListOf<Long>()
         input.forEach {
-            left.add(it.substringBefore(" ").toInt())
-            right.add(it.substringAfterLast(" ").toInt())
+            left.add(it.substringBefore(" ").toLong())
+            right.add(it.substringAfterLast(" ").toLong())
         }
         return left to right
     }
